Revert "Refactor routers into individual modules"
This reverts commit 101602d72c89c530be2e563fe067d8190bc60891.
