+++ services/ad-tech/src/main.ts
@@ -13,20 +13,12 @@
 
 import express, {Application} from 'express';
 import cors from 'cors';
+
 import {HOSTNAME, PORT} from './lib/constants.js';
-import {AdsRouter} from './routes/common/ads-router.js';
-import {AttributionReportingRouter} from './routes/common/attribution-reporting-router.js';
-import {CommonRouter} from './routes/common/common-router.js';
-import {ReportRouter} from './routes/common/report-router.js';
-import {TopicsRouter} from './routes/common/topics-router.js';
-import {BiddingSignalsRouter} from './routes/dsp/bidding-signals-router.js';
-import {BuyerContextualBidderRouter} from './routes/dsp/buyer-contextual-bidder.js';
-import {BuyerRouter as BuyerMiscRouter} from './routes/dsp/buyer-router.js';
-import {ScoringSignalsRouter} from './routes/ssp/scoring-signals-router.js';
-import {SellerContextualBidder} from './routes/ssp/seller-contextual-bidder.js';
-import {SellerRouter} from './routes/ssp/seller-router.js';
-import {WellKnownAttributionReportingRouter} from './routes/well-known/attribution-reporting-router.js';
-import {WellKnownPrivateAggregationRouter} from './routes/well-known/private-aggregation-router.js';
+import {WellKnownRouter} from './routes/well-known-router.js';
+import {BuyerRouter} from './routes/buyer-router.js';
+import {CommonRouter} from './routes/common-router.js';
+import {SellerRouter} from './routes/seller-router.js';
 
 const app: Application = express();
 app.use(express.urlencoded({extended: true}));
@@ -36,25 +28,9 @@ app.set('view engine', 'ejs');
 app.set('views', 'src/views');
 
 app.use('/', CommonRouter);
-
-app.use('/.well-known/private-aggregation', WellKnownPrivateAggregationRouter);
-app.use(
-  '/.well-known/attribution-reporting',
-  WellKnownAttributionReportingRouter,
-);
-
-app.use('/ads', AdsRouter);
-app.use('/reporting', ReportRouter);
-app.use('/attribution', AttributionReportingRouter);
-app.use('/topics', TopicsRouter);
-
-app.use('/dsp', BuyerMiscRouter);
-app.use('/dsp/contextual-bid', BuyerContextualBidderRouter);
-app.use('/dsp/realtime-signals', BiddingSignalsRouter);
-
+app.use('/.well-known', WellKnownRouter);
+app.use('/dsp', BuyerRouter);
 app.use('/ssp', SellerRouter);
-app.use('/ssp/contextual-bid', SellerContextualBidder);
-app.use('/ssp/realtime-signals', ScoringSignalsRouter);
 
 app.listen(PORT, function () {
   console.log(HOSTNAME, ' listening on port ', PORT);
